pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "panelflow"
version = "0.1.0"
description = "AI-powered comic panel video content creator"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "opencv-python",
    "librosa",
    "pillow",
    "google-auth",
    "google-auth-oauthlib",
    "google-api-python-client",
    "ollama",
    "pyautogui",
    "requests-oauthlib",
    "cairosvg",
    "lxml",
    "pydub",
    "moviepy==1.0.3",
    "fastapi",
    "uvicorn",
    "python-multipart",
    "websockets",
    "pytesseract",
    "triton",
    "langdetect",
    "pysubs2",
    "hmmlearn",
    "eyed3",
    "imbalanced-learn",
    "ffmpeg-python",
    "scenedetect",
    "pyannote.audio",
    "google-generativeai",
    "rarfile",
    "ffmpeg",
    "pymediainfo",
    "beautifulsoup4",
    "piexif",
    "jebin_lib @ git+https://github.com/jebin2/lib.git",
    "custom_logger @ git+https://github.com/jebin2/custom_logger.git",
    "gemiwrap @ git+https://github.com/jebin2/gemiwrap.git",
    "browser_manager @ git+https://github.com/jebin2/browser_manager.git",
    "chat_bot_ui_handler @ git+https://github.com/jebin2/chat_bot_ui_handler.git",
    "youtube_auto_pub @ git+https://github.com/jebin2/youtube_auto_pub.git",
    "psutil",
    "svgwrite",
    "spacy",
    "json_repair",
    "pydantic",
    "imagehash",
    "huggingface_hub",
    "cryptography",
    "sentence_transformers",
    "easyocr",
    "chardet",
]

[project.scripts]
panelflow = "main:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
panelflow = ["assets/**", "prompt/**"]
